backend/requirements.txt
Django
gunicorn
psycopg2-binary
cryptography
djangorestframework
django-filter
django-cors-headers
drf-yasg
django-allauth
django-url-filter
django-safedelete
bcrypt
python-dotenv
python-dateutil
rest-framework-roles
django-rest-swagger
dj-rest-auth
Pillow
deepface
django-cleanup
channels
channels-redis
asgi-redis
redis